work-time/requirements.txt

python-telegram-bot==20.0
python-dotenv
opencv-python
pytesseract
aiohttp
